+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [1.4.10] - 2022-09-
+## [1.4.10] - 2022-09-30
 
 ### Added
 - The `signature pen thickness scaling factor` Configuration directive.
@@ -8,6 +8,12 @@
 - Upgraded Mako to version 1.2.2.
 - The field modifiers `address autocomplete`, `label above field`, and
   `floating label` can now refer to Python expressions.
+- Upgraded Google OAuth.
+### Fixed
+- Issue with exim4's overinclusive `dc_other_hostnames` setting.
+- Error with images in Markdown to DOCX conversion.
+- Error with `overlay_pdf()`.
+- Error with `temporary session` when user is not logged in.
 
 ## [1.4.9] - 2022-09-10